fix(Enabling Bullet Plot): Refactor of the bullet plot to use the parent and child values rather than calculating the average. Some other plot improvements as well
Along with the parent and child values, the tooltips and legend text now reflect the regions which they are showing. It also shows all bars even where data is missing. It changes nodata value from 0 to null as 0 is a meaningful value in this plot (low adaptive capacity). Sorting is alphabetical to allow easier comaprisions between plots.