feat(data): Add a data file for countries to allow easier development, access to translations
Should be used for selectors to include the given countries in the atlas. This includes the Iso3c code, allows a toggle to easily add or remove countries based on the boundaries we have. Hopefully will allow better load tome as select elements will not need to rely on the data to get an initial list of countries.