fix(ROI Tool): updates to styling so that tooltips can expand out onto the page and hiding some narrative in an optional dropdown
Update to reflect changes to global css allowing cells to expand into rest of body. This broke the width of the country/crop/region select. Hiding some narrative as to not overwhem reader and allow them to get to the fun interactive bits faster

BREAKING CHANGE: Broke the width do the top select